# Conflicts:
#	backend/src/Squidex.Domain.Apps.Entities.MongoDb/Assets/MongoAssetRepository.cs
#	backend/src/Squidex.Domain.Apps.Entities.MongoDb/Assets/MongoAssetRepository_SnapshotStore.cs
#	backend/src/Squidex.Infrastructure.MongoDb/MongoDb/BsonDomainIdSerializer.cs
#	backend/src/Squidex/Config/Domain/StoreServices.cs

+++ backend/src/Squidex.Infrastructure/EventSourcing/EnvelopeExtensions.cs
@@ -32,7 +32,7 @@ public static class EnvelopeExtensions
 
     public static Envelope<T> SetEventStreamNumber<T>(this Envelope<T> envelope, long value) where T : class, IEvent
     {
-        envelope.Headers[CommonHeaders.EventStreamNumber] = (double)value;
+        envelope.Headers[CommonHeaders.EventStreamNumber] = value;
 
         return envelope;
     }
